refactor: bump 0.4.0
