+++ Cargo.toml
@@ -1,29 +1,25 @@
 [package]
 name = "nom-rule"
-version = "0.3.1"
+version = "0.4.0"
 authors = ["andylokandy"]
 edition = "2021"
 license = "MIT"
-description = "A procedural macro for defining nom combinators in simple DSL"
+description = "A procedural macro for writing nom parsers using a grammar-like DSL."
 homepage = "https://github.com/andylokandy/nom-rule"
 repository = "https://github.com/andylokandy/nom-rule"
-documentation = "https://docs.rs/nom-rule"
 readme = "README.md"
-keywords = ["nom", "parsing", "generator", "pest", "macro"]
+keywords = ["nom", "parser", "grammar", "DSL", "parsing"]
 
 [lib]
 proc-macro = true
 
 [dependencies]
 nom = "7"
-syn = "1"
-quote = "1"
-pratt = "0.3"
-proc-macro2 = { version = "1", features = ["span-locations"] }
+pratt = "0.4"
 proc-macro-error = "1"
+proc-macro2 = { version = "1", features = ["span-locations"] }
+quote = "1"
+syn = "2"
 
 [dev-dependencies]
-logos = "0.12"
-
-[features]
-auto-sequence = []
+logos = "0.14"
